test: #6557 마킹 회귀 테스트를 src 단위 테스트에서 tests/cases 로 옮긴다
CI 정책 게이트가 PR base 대비 source-side(#[cfg(test)]) 테스트 총량
증가를 금지한다 (4223 > 4221 로 Lint 실패). 새 회귀 테스트 2건을
tests/cases/issue_6557_merged_cell_boundary_drag_marking.rs 로 옮기고
공개 API(HwpDocument::resize_table_cells / merge_table_cells_native)
경계에서 같은 계약을 검증한다. 동작 코드는 변화 없음.
