+++ src/wasm_api/tests.rs
@@ -28565,108 +28565,3 @@ fn issue_4576_rebuild_derived_state_recomputes_composition_and_pagination() {
         "페이지 트리가 기준선과 같아야 한다 (옛 코드 잔재 없음)"
     );
 }
-
-#[test]
-fn cell_selection_boundary_drag_with_vertical_merge_keeps_uniform_grid() {
-    // 3x2 표에서 (0,0)-(1,0) 세로 병합 후, 셀 선택 열 경계 드래그가 보내는
-    // 폭 보존 update(경계 왼쪽 셀 +d, 오른쪽 이웃 -d)를 적용한다. 결과가 전 행
-    // 균일하면 local_resize 행을 마킹하지 않아야 한다 — 병합 셀의 delta 는
-    // 홈 행(row0)에만 집계되므로, 종전 count/합 휴리스틱은 row0·row2 만 마킹했고
-    // base grid 추출(local 행 제외, col_span==1 max)에서 col0 이 행 폭 소스를
-    // 잃어 기본값 1800 으로 무너졌다 (스튜디오 실측: 병합 셀 279.7px → 24px,
-    // 잔여 폭은 마지막 열으로 쏠림).
-    let mut doc = HwpDocument::create_empty();
-    let created = doc.create_table_native(0, 0, 0, 3, 2).expect("3x2 표");
-    let para_idx = issue_1481_json_usize(&created, "paraIdx");
-    doc.merge_table_cells_native(0, para_idx, 0, 0, 0, 1, 0)
-        .expect("세로 병합");
-
-    let find_cell = |doc: &HwpDocument, row: u16, col: u16| -> usize {
-        issue_1481_table(doc, para_idx)
-            .cells
-            .iter()
-            .enumerate()
-            .find(|(_, c)| c.row == row && c.col == col)
-            .map(|(idx, _)| idx)
-            .expect("셀 인덱스")
-    };
-    let merged_idx = find_cell(&doc, 0, 0);
-    let r0c1 = find_cell(&doc, 0, 1);
-    let r1c1 = find_cell(&doc, 1, 1);
-    let r2c0 = find_cell(&doc, 2, 0);
-    let r2c1 = find_cell(&doc, 2, 1);
-    let (base_w0, base_w1) = {
-        let table = issue_1481_table(&doc, para_idx);
-        (table.cells[merged_idx].width, table.cells[r0c1].width)
-    };
-
-    let delta = 3570i32;
-    let neg = -delta;
-    let payload = format!(
-        r#"[{{"cellIdx":{merged_idx},"widthDelta":{delta}}},{{"cellIdx":{r0c1},"widthDelta":{neg}}},{{"cellIdx":{r1c1},"widthDelta":{neg}}},{{"cellIdx":{r2c0},"widthDelta":{delta}}},{{"cellIdx":{r2c1},"widthDelta":{neg}}}]"#
-    );
-    doc.resize_table_cells_native(0, para_idx, 0, &payload)
-        .expect("경계 드래그 적용");
-
-    let table = issue_1481_table(&doc, para_idx);
-    assert!(
-        table.local_resize_rows.is_empty(),
-        "전 행 균일 결과는 행 단위 resize 마킹 대상이 아니다: {:?}",
-        table.local_resize_rows
-    );
-    let widths = table.get_column_widths();
-    assert_eq!(
-        widths[0] as i64,
-        base_w0 as i64 + delta as i64,
-        "col0 폭 = 원래 + delta"
-    );
-    assert_eq!(
-        widths[1] as i64,
-        base_w1 as i64 - delta as i64,
-        "col1 폭 = 원래 - delta"
-    );
-    for cell in &table.cells {
-        if cell.col_span == 1 {
-            assert_eq!(
-                cell.width, widths[cell.col as usize],
-                "모든 행이 base grid 와 같은 폭을 가져야 한다 (row{} col{})",
-                cell.row, cell.col
-            );
-        }
-    }
-}
-
-#[test]
-fn cell_selection_boundary_drag_divergent_row_still_marks_local_resize() {
-    // 위 테스트의 대조군: 결과가 실제로 갈라지면(선택 행만 이동) 종전과 같이
-    // 그 행을 local_resize 로 마킹해야 한다. 3x2 병합 없는 표에서 row1 만
-    // +d/-d 를 받으면 row1 의 폭 벡터가 base grid 와 달라진다.
-    let mut doc = HwpDocument::create_empty();
-    let created = doc.create_table_native(0, 0, 0, 3, 2).expect("3x2 표");
-    let para_idx = issue_1481_json_usize(&created, "paraIdx");
-
-    let find_cell = |doc: &HwpDocument, row: u16, col: u16| -> usize {
-        issue_1481_table(doc, para_idx)
-            .cells
-            .iter()
-            .enumerate()
-            .find(|(_, c)| c.row == row && c.col == col)
-            .map(|(idx, _)| idx)
-            .expect("셀 인덱스")
-    };
-    let r1c0 = find_cell(&doc, 1, 0);
-    let r1c1 = find_cell(&doc, 1, 1);
-
-    let payload = format!(
-        r#"[{{"cellIdx":{r1c0},"widthDelta":900}},{{"cellIdx":{r1c1},"widthDelta":-900}}]"#
-    );
-    doc.resize_table_cells_native(0, para_idx, 0, &payload)
-        .expect("행 한정 드래그 적용");
-
-    let table = issue_1481_table(&doc, para_idx);
-    assert_eq!(
-        table.local_resize_rows,
-        vec![1],
-        "갈라진 행은 종전대로 행 단위 resize 로 마킹된다"
-    );
-}
